Add auth.go and user.go models
