+++ domain/models/user.go
@@ -1,7 +1,6 @@
-package domain
+package models
 
 import (
-	"context"
 	"time"
 )
 
@@ -51,18 +50,4 @@ type User struct {
 	UpdatedAt         time.Time
 
 	Provider          string
-}
-type IEmailService interface {
-	SendEmail(to, subject, body string) error
-}
-
-type IUserRepository interface {
-	// GetProfile(ctx context.Context) (*User, error)
-	UpdateProfile(ctx context.Context, user *User) (*User, error)
-	GetByID(ctx context.Context, id string) (*User, error)
-}
-
-type IUserUsecase interface {
-	UpdateProfile(ctx context.Context, user *User) (*User, error)
-	GetProfile(ctx context.Context, userID string) (*User, error)
 }
